[PRJ][APPS][USBH] Fix stack corruption due to inconsistent type of local variable in MSC_File_Operations()

+++ Projects/STM32469I-Discovery/Applications/USB_Host/FWupgrade_Standalone/Src/command.c
@@ -190,7 +190,7 @@ static void COMMAND_ProgramFlashMemory(void)
 {
   uint32_t programcounter = 0x00;
   uint8_t readflag = TRUE;
-  uint16_t bytesread;
+  uint32_t bytesread;
 
   /* RAM Address Initialization */
   RamAddress = (uint32_t) &RAM_Buf;
